tidy(zoho): Split live tests
